Fixed data length calculation to include +1 offset

+++ inc/MessageParser.hpp
@@ -98,7 +98,7 @@ private:
 	 * @param length The size of the header
 	 * @param message The Message to modify based on the header
 	 */
-	static void parseECSSTCHeader(const uint8_t* data, uint16_t length, Message& message);
+	static void parseECSSTCHeader(const uint8_t* data, uint32_t length, Message& message);
 
 	/**
 	 * Parse the ECSS Telemetry packet secondary header
@@ -109,7 +109,7 @@ private:
 	 * @param length The size of the header
 	 * @param message The Message to modify based on the header
 	 */
-	static void parseECSSTMHeader(const uint8_t* data, uint16_t length, Message& message);
+	static void parseECSSTMHeader(const uint8_t* data, uint32_t length, Message& message);
 };
 
 #endif // ECSS_SERVICES_MESSAGEPARSER_HPP
